Modified Traditional African and Caribbean Recipes
We are a team of nutrition academics, Dr Tanefa Apekey (Registered Nutritionist; University of Sheffield) and Dr Sally Moore (Registered Dietitian & Nutritionist; University of Leeds), who are responding to the need for nutrition professionals to access evidence-based culturally appropriate healthy eating and recipe resources to guide food choice with people from certain ethnic groups, including people from African and Caribbean backgrounds. To do this, we have worked in partnership with people from the relevant ethnic groups and a community partner based in Leeds Chapeltown, to co-design and test healthier recipes featuring African and Caribbean traditional dishes. This is an ongoing project and we will continue to update this page with more recipes.